Address PR comments, rename isTarImport param to be cleanup_on_failure

+++ app/controllers/assessments_controller.rb
@@ -115,7 +115,7 @@ class AssessmentsController < ApplicationController
         # add line break if adding to existing error message
         flash.now[:error] = flash.now[:error] ? "#{flash.now[:error]} <br>" : ""
         flash.now[:error] += "An error occurred while trying to display an existing assessment " \
-            "on file directory #{filename}: assessment file names must only contain lowercase " \
+            "from file directory #{filename}: assessment file names must only contain lowercase " \
             "letters and digits with no spaces"
         flash.now[:html_safe] = true
         next
@@ -129,8 +129,8 @@ class AssessmentsController < ApplicationController
                                ))
         unless props["general"] && (props["general"]["name"] == filename)
           flash.now[:error] = flash.now[:error] ? "#{flash.now[:error]} <br>" : ""
-          flash.now[:error] += "An error occurred while trying to display an existing assessment" \
-          "on file directory #{filename}: Name in yaml (#{props['general']['name']}) " \
+          flash.now[:error] += "An error occurred while trying to display an existing assessment " \
+          "from file directory #{filename}: Name in yaml (#{props['general']['name']}) " \
           "doesn't match #{filename}"
           flash.now[:html_safe] = true
           next
@@ -138,7 +138,7 @@ class AssessmentsController < ApplicationController
       else
         flash.now[:error] = flash.now[:error] ? "#{flash.now[:error]} <br>" : ""
         flash.now[:error] += "An error occurred while trying to display an existing assessment " \
-          "on file directory #{filename}: #{filename}.yml does not exist"
+          "from file directory #{filename}: #{filename}.yml does not exist"
         flash.now[:html_safe] = true
         next
       end
@@ -232,7 +232,7 @@ class AssessmentsController < ApplicationController
     end
 
     params[:assessment_name] = asmt_name
-    params[:isTarImport] = true
+    params[:cleanup_on_failure] = true
     importAssessment && return
   end
 
@@ -242,7 +242,7 @@ class AssessmentsController < ApplicationController
   action_auth_level :importAssessment, :instructor
 
   def importAssessment
-    isTarImport = params[:isTarImport]
+    cleanup_on_failure = params[:cleanup_on_failure]
     @assessment = @course.assessments.new(name: params[:assessment_name])
     assessment_path = Rails.root.join("courses/#{@course.name}/#{@assessment.name}")
     # not sure if this check is 100% necessary anymore, but is a last resort
@@ -254,9 +254,7 @@ class AssessmentsController < ApplicationController
       destroy_no_redirect
       # delete files explicitly b/c the paths don't match ONLY if
       # import was from tarball
-      if isTarImport
-        FileUtils.rm_rf(assessment_path)
-      end
+      FileUtils.rm_rf(assessment_path) if cleanup_on_failure
       redirect_to(install_assessment_course_assessments_path(@course)) && return
     end
 
@@ -267,9 +265,7 @@ class AssessmentsController < ApplicationController
       destroy_no_redirect
       # delete files explicitly b/c the paths don't match ONLY if
       # import was from tarball
-      if isTarImport
-        FileUtils.rm_rf(assessment_path)
-      end
+      FileUtils.rm_rf(assessment_path) if cleanup_on_failure
       redirect_to(install_assessment_course_assessments_path(@course)) && return
     end
     @assessment.load_embedded_quiz # this will check and load embedded quiz
@@ -281,9 +277,7 @@ class AssessmentsController < ApplicationController
       destroy_no_redirect
       # delete files explicitly b/c the paths don't match ONLY if
       # import was from tarball
-      if isTarImport
-        FileUtils.rm_rf(assessment_path)
-      end
+      FileUtils.rm_rf(assessment_path) if cleanup_on_failure
       redirect_to(install_assessment_course_assessments_path(@course)) && return
     end
     flash[:success] = "Successfully imported #{@assessment.name}"
